+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # next
 - little improvement in `drawDefaultLifeBar`.
 - create mixin `AutomaticRandomMovement`
+- add `onReady` in `BonfireTiledWidget` and `BonfireWidget`
 
 # 1.3.0
 - new extensions to `GameComponent`.